test(encoders): Add dtypes to encoder tests

+++ tests/data/encoding/test_encoders.py
@@ -248,7 +248,7 @@ class TestTextAsciiEncoder:
 
     def test_encode_dtype(self) -> None:
         """Test encoding with a non-default dtype."""
-        encoder = TextAsciiEncoder(dtype="int32")
+        encoder = TextAsciiEncoder(dtype=torch.int32)
         input_str = "hello"
         output = encoder.encode(input_str)
         assert output.dtype == torch.int32
@@ -438,7 +438,7 @@ class TestStrClassificationEncoder:
         Returns:
             StrClassificationEncoder: Default encoder instance
         """
-        return StrClassificationEncoder()
+        return StrClassificationEncoder(dtype=torch.int64)
 
     @staticmethod
     @pytest.fixture
@@ -448,7 +448,7 @@ class TestStrClassificationEncoder:
         Returns:
             StrClassificationEncoder: Scaled encoder instance
         """
-        return StrClassificationEncoder(scale=True)
+        return StrClassificationEncoder(scale=True, dtype=torch.float32)
 
     @pytest.mark.parametrize("fixture", ["str_encoder", "scaled_encoder"])
     def test_encode_raises_not_implemented(
@@ -544,7 +544,7 @@ class TestNumericRankEncoder:
         Returns:
             NumericRankEncoder: Scaled encoder instance
         """
-        return NumericRankEncoder(scale=True)
+        return NumericRankEncoder(scale=True, dtype=torch.float32)
 
     @pytest.mark.parametrize("fixture", ["rank_encoder", "scaled_encoder"])
     def test_encode_raises_not_implemented(
